requirements.txt
absl-py==1.4.0
appdirs==1.4.4
array-record==0.2.0
astunparse==1.6.3
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_07078715-3ab7-4562-8d3d-d56b0eaa0f7dp504n_ny/croots/recipe/bottleneck_1657175566567/work
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ab19r4bji3/croot/cffi_1670423206034/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a9zjqvif1f/croot/cryptography_1677533099634/work
cvxopt==1.2.6
cvxpy==1.3.1
dm-tree==0.1.8
ecos==2.0.12
etils==1.2.0
exceptiongroup==1.1.1
fancyimpute==0.7.0
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
h5py @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/h5py_1637162753555/work
idna @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_771olrhiqw/croot/idna_1666125579282/work
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
jax==0.4.8
Jinja2==3.1.2
joblib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_70fph11sxu/croot/joblib_1666298846160/work
keras==2.12.0
knnimpute==0.1.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
missingpy==0.2.0
ml-dtypes==0.1.0
nose==1.3.7
numexpr @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_76yyu1p9jk/croot/numexpr_1683221830860/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1660672958449/work
oauthlib==3.2.2
opt-einsum==3.3.0
osqp==0.6.2.post9
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e946luvhc3/croot/packaging_1678965323926/work
pandas==1.5.3
pluggy==1.0.0
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
promise==2.3
protobuf==4.22.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_38h4axtq38/croot/pyopenssl_1677607699670/work
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
pytest==7.3.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qdldl==0.1.7
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_37wt8ycste/croot/requests_1682607533309/work
requests-oauthlib==1.3.1
rpy2==3.4.5
rsa==4.9
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1659726157499/work
scikit-umfpack==0.3.3
scipy==1.10.1
scs==3.2.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
tensorflow-metadata==1.13.1
tensorflow-metal==0.8.0
termcolor==2.3.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_c7xh0_scrc/croot/urllib3_1680254693282/work
Werkzeug==2.3.3
wrapt==1.14.1
zipp==3.15.0
